fixed object ref not set to instance of object
botName wasn't being grabbed successfully in MessageReceived. I forgot to remove this old line of code in the public build my bad.

The BotName is now gotten via MainLoop.cs at login. It tries to use the bot's nickname if it has one, but gets the username instead if not.