updated text-gen-webui instructions for new API changes
oobabooga changed to openAPI format for the API and silently deleted the old API that sallybot uses. I updated instructions to show how to quickly copy over the old API folder so you can be back up and running.

Specifically this change is the stupid image that didn't update somehow? I pasted it, so idk how it has the same image. whatever, I double checked, this link IS the new image, so if it's busted then github is broken